+++ examples/basic-example/index.js
@@ -6,7 +6,7 @@ const fromPySyft = `
     [1,
      [6,
       [[6,
-        [[5, [b'__add__']],
+        [[5, [b'__add__',]],
          [18, [71560559247, 2472411137, 28099456635, None, [2], False]],
          [6, [[18, [71560559247, 2472411137, 28099456635, None, [2], False]]]],
          [0, []]]],
@@ -15,7 +15,7 @@ const fromPySyft = `
     [1,
      [6,
       [[6,
-        [[5, [b'torch.abs']],
+        [[5, [b'torch.abs',]],
          None,
          [6,
           [[18, [32408443280, 99020234254, 28099456635, None, None, True]]]],
